% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioplot.R
\name{audioplot}
\alias{audioplot}
\title{Fast Spectrogram with Derived Pseudo-Oscillogram and Power Spectrum}
\usage{
audioplot(
  wave,
  fft = "auto",
  enhance = "time",
  overlap = 75,
  window = "hanning",
  contrast = 50,
  flim = NULL,
  tlim = NULL,
  zero.pad = 8,
  output.file = NULL,
  plot.title = "",
  use.ragg = TRUE,
  palette = "Mako",
  palette_rev = FALSE,
  osc.show = TRUE,
  osc.fun = "mean",
  osc.color = "black",
  osc.height = 0.25,
  spec.show = TRUE,
  spec.fun = "mean",
  spec.color = "black",
  spec.width = 0.15,
  spec.linear = TRUE
)
}
\arguments{
\item{wave}{A Wave object created with tuneR package (e.g., readWave).}

\item{fft}{Numeric. Size of FFT window (must be power of 2). Default is
"auto", which automatically calculates a proper length based on 'enhance'.}

\item{enhance}{Character. Either 'freq' to prioritize frequency
resolution or 'time' to prioritize temporal resolution when fft="auto".}

\item{overlap}{Numeric. Overlap between windows (0 to 95\%)}

\item{window}{Character. Type of window function. Default is "hanning".
Also accepts "hamming" and "blackman".}

\item{contrast}{Numeric. Dynamic range for visualization (0-100\%).}

\item{flim}{Numeric vector of length 2. Frequency limits, in kHz.}

\item{tlim}{Numeric vector of length 2. Time limits, in seconds.}

\item{zero.pad}{Numeric. Zero-padding factor.}

\item{output.file}{Character. Optional file path to save the spectrogram.}

\item{plot.title}{Character. Plot title}

\item{use.ragg}{Logical. Whether to use RAGG for GPU-accelerated rendering.}

\item{palette}{Character. Color palette for spectrogram. Can be a
character string (palette name from \code{hcl.colors}) or a function
that returns a color vector.}

\item{palette_rev}{Logical. Whether to reverse the color palette. To obtain
a white background with dark signals, set to TRUE and use palette="Grays".}

\item{osc.show}{Logical. Whether to show the pseudo-oscillogram}

\item{osc.fun}{Character. Name of the summary statistic used to create
the pseudo-oscillogram (e.g., "mean", "max", "sd", etc.)}

\item{osc.color}{Character. Color for the pseudo-oscillogram}

\item{osc.height}{Numeric. Relative height of oscillogram (0-1)}

\item{spec.show}{Logical. Whether to show the power spectrum}

\item{spec.fun}{Character. Summary function to apply across frequency
bins for power spectrum calculation (e.g., "mean", "max", "sd", etc.).}

\item{spec.color}{Characer. Color for the power spectrum.}

\item{spec.width}{Numeric. Relative width of power spectrum (0-1).}

\item{spec.linear}{Logical. Whether to show the power spectrum in linear scale. Default
is TRUE.}
}
\value{
Invisible list containing spectrogram data
}
\description{
Fast Spectrogram with Derived Pseudo-Oscillogram and Power Spectrum
}
\details{
When fft is set to 'auto' (recommended), the function takes the 'enhance'
argument to decide whether to enhance spectral or temporal resolution. We
recommend "enhance= 'freq'" for high-quality factor (i.e., 'tonal') sounds
such as those produced by most crickets, and "enhance='time" for low-quality
factor (i.e., 'raspy') sounds such as those produced by katydids and cicadas.
}
\examples{
\dontrun{
library(RthopteraSounds)
data("coryphoda")
data("gryllus")
audioplot(coryphoda, enhance = "time",
osc.show = FALSE, palette = 'Zissou 1')
audioplot(coryphoda, enhance = 'time', osc.show = FALSE,
spec.linear = TRUE) # great
audioplot(coryphoda, enhance = "time")
audioplot(gryllus, enhance = "freq")
audioplot(gryllus, enhance = "time")

# Grayscale with white background
audioplot(coryphoda, palette = "Grays", palette_rev = TRUE)
}
}
