% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_stats_hq.R
\name{call_stats_hq}
\alias{call_stats_hq}
\title{Temporal and spectral statistics for HQ calls}
\usage{
call_stats_hq(
  wave,
  specimen_id = "",
  msmooth_window = 100,
  msmooth_overlap = 50,
  upper_detection_threshold = 0.2,
  lower_detection_threshold = 0.1,
  min_train_dur = 0.002,
  max_train_gap = 0.08,
  motif_seq = TRUE,
  max_motif_gap = 0.8,
  norm_env = TRUE,
  db_threshold = 20
)
}
\arguments{
\item{wave}{A Wave object containing the sound data.}

\item{specimen_id}{A character string representing the specimen ID.}

\item{msmooth_window}{An integer specifying the window size (in milliseconds)
for smoothing the envelope (default: 100).}

\item{msmooth_overlap}{An integer specifying the overlap (in %) for smoothing
the envelope. Default = 50.}

\item{upper_detection_threshold}{A numeric value representing the upper
amplitude detection threshold Default = 0.2.}

\item{lower_detection_threshold}{A numeric value representing the lower
amplitude detection threshold Default = 0.1.}

\item{min_train_dur}{A numeric value specifying the minimum duration (in
seconds) of a valid amplitude train. Default = 0.002.}

\item{max_train_gap}{A numeric value representing the maximum gap (in
seconds) allowed between two trains before considering them part of different
motifs. Default = 0.08.}

\item{motif_seq}{Logical. If TRUE, add an first-order aggregation of motifs. Default
value is FALSE.}

\item{max_motif_gap}{Numeric. Maximum allowed gap (in seconds) between motifs
to consider them part of the same group.
 @param norm_env A logical indicating whether to normalize the envelope
 Default = TRUE.}
}
\value{
A list with the following components:
\item{plot}{An interactive `plotly` plot showing the signal envelope,
detected trains, and motifs.}
\item{summary_data}{A tibble with summary statistics of the acoustic motifs.}
\item{train_data}{A tibble with details about the detected trains.}
\item{motif_data}{A tibble with details about the detected motifs.}
\item{params}{A tibble containing the input parameters.}
}
\description{
This function analyzes the acoustic characteristics of a wave object by
detecting sounds above a threshold in the envelope created with a mean
smoothing window, grouping them into motifs, and computing various metrics
related to their temporal and spectral structure.
An interactive plot visualizes the results, showing the envelope of the
signal, detected trains, motifs, and key statistics. The plot can be
downloaded as an HTML (interactive) or PNG (static) file.
}
\examples{
\dontrun{
data(gryllus)
# result <- temporal_stats_hq(gryllus)
# result$plot
}
}
