% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wave.R
\name{clean_wave}
\alias{clean_wave}
\title{Denoise a Wave for temporal measurements}
\usage{
clean_wave(
  wave,
  filter = "fir",
  min = NULL,
  max = NULL,
  gate = 0.05,
  oscillo = TRUE
)
}
\arguments{
\item{wave}{a Wave object.}

\item{filter}{Character. Select a bandpass filter. Options are 'fir' (Finite
Impulse Response);'bwf' (Butterworth); and 'freq' (Frequency filter).}

\item{min}{Numeric. The minimum frequency for the bandpass filter.}

\item{max}{Numeric. The maximum frequency for the bandpass filter.}

\item{gate}{Numeric (double). The proportion of the maximum amplitude to use
as a threshold for the noise gate. The values below this threshold are
clipped to zero.}

\item{oscillo}{Logical. If TRUE (default), a simple oscillogram of the
output wave is plotted.}
}
\value{
a Wave object
}
\description{
Denoise a Wave for temporal measurements
}
\details{
This function was developed as a preprocessing step for temporal measurements
only. DO NOT conduct spectral analysis on objects returned by this function.
}
\examples{
\dontrun{
library(RthopteraSounds)
data(Coryphoda)
clean_wave(Coryphoda)
}
}
