% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_app.R
\name{downsample_app}
\alias{downsample_app}
\title{Downsample Audio Shiny App}
\usage{
downsample_app()
}
\arguments{
\item{input}{Internal Shiny parameter for handling input values.}

\item{output}{Internal Shiny parameter for rendering outputs in the app.}

\item{session}{Internal Shiny parameter for managing the app session.}
}
\value{
A Shiny app interface for downsampling and visualizing audio wave objects.
}
\description{
This function launches a Shiny app that allows users to downsample audio (wave objects) and visualize the mean spectrum. Users can select an existing wave object from the R environment, specify a new maximum frequency (kHz), downsample the audio, and save the downsampled wave object with a new name.
}
\examples{
\dontrun{
downsample_app()
}
}
