% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcidectes.R
\docType{data}
\name{falcidectes}
\alias{falcidectes}
\title{Recording of Falcidectes divisus}
\format{
An object of class \code{Wave} of length 411098.
}
\source{
Original recording by Francisco Rivas Fuenzalida.
}
\usage{
falcidectes
}
\description{
The calling song of *Falcidectes sp.*, a Tettigoniidae from Central Chile.
}
\details{
Sampling rate: 96 kHz.

Bit depth: 16 bit.

Duration: 4.3 sec.
}
\examples{
\dontrun{
data(coryphoda22C)
summary(coryphoda22C)
spectrum_plotly(coryphoda22C)
}
}
\keyword{datasets}
