% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gryllus.R
\docType{data}
\name{gryllus}
\alias{gryllus}
\title{Recording of Gryllus campestris}
\format{
An object of class \code{Wave} of length 217900.
}
\source{
\url{https://xeno-canto.org/753101}.
}
\usage{
gryllus
}
\description{
The calling song of a *Gryllus campestris* recorded in Italy by Cesare Brizio on July 6th, 2006.
}
\details{
Sampling rate: 44.1 kHz.

Snippet duration: ~5 seconds.
}
\examples{
\dontrun{
data(gryllus)
summary(gryllus)
spectrum_plotly(gryllus)
}
}
\keyword{datasets}
