% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoplosphyrum.R
\docType{data}
\name{hoplosphyrum}
\alias{hoplosphyrum}
\title{Recording of Hoplosphyrum griseum}
\format{
An object of class \code{Wave} of length 1078355.
}
\source{
Original recording by Francisco Rivas Fuenzalida.
}
\usage{
hoplosphyrum
}
\description{
The calling song of *Hoplosphyrum griseum*, a Mogoplistidae from Central Chile.
}
\details{
Sampling rate: 48 kHz.

Bit depth: 16 bit.

Duration: ~23 sec.
}
\examples{
\dontrun{
data(hoplosphyrum)
summary(hoplosphyrum)
spectrum_plotly(hoplosphyrum)
}
}
\keyword{datasets}
