% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch a Shiny App from the package}
\usage{
launchApp(app_name = NULL)
}
\arguments{
\item{app_name}{The (partial or full) name of the Shiny app folder inside inst/shiny. If NULL, lists available apps.}
}
\value{
Launches the Shiny app or lists available apps if no name is provided
}
\description{
Launch a Shiny App from the package
}
