% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rthoptera_data.R
\name{load_rthoptera_data}
\alias{load_rthoptera_data}
\title{Load all data items in Rthoptera}
\usage{
load_rthoptera_data()
}
\value{
All the Wave objects included in the package will be available in the environment.
}
\description{
Load all data items in Rthoptera
}
\examples{
\dontrun{
load_rthoptera_data()
}
}
