% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanogryllus.R
\docType{data}
\name{melanogryllus}
\alias{melanogryllus}
\title{Recording of Melanogryllus desertus}
\format{
An object of class \code{Wave} of length 2500000.
}
\source{
Original recording by Cesare Brizio.
}
\usage{
melanogryllus
}
\description{
A single calling song of *Melanogryllus desertus*, a Gryllidae recorded
by Cesare Brizio in Italy.
}
\details{
Sampling rate: 250 kHz.

Bit depth: 16 bit.

Duration: 10 sec.
}
\examples{
\dontrun{
data(melanogryllus)
summary(melanogryllus)
spectrum_plotly(melanogryllus)
}
}
\keyword{datasets}
