% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_meanspectra_static_app.R
\name{multi_meanspectra_static_app}
\alias{multi_meanspectra_static_app}
\title{Multi-Power Spectra Shiny App}
\usage{
multi_meanspectra_static_app()
}
\value{
A Shiny app for visualizing wave oscillograms and mean spectra.
}
\description{
This Shiny app allows users to visualize and interact with the oscillogram and power spectra of selected wave objects. Users can add multiple selections on the oscillogram and visualize the corresponding mean power spectra. The app provides options to download individual or combined visualizations.
}
\examples{
if (interactive()) {
  multi_meanspectra()
}
}
