% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioscillo_app.R
\name{multioscillo_app}
\alias{multioscillo_app}
\title{Multi-Oscillogram Shiny App}
\usage{
multioscillo_app(dark.mode = TRUE)
}
\arguments{
\item{dark.mode}{Logical, indicating whether the app should use a dark mode theme (default is TRUE).}
}
\value{
A Shiny app interface for plotting and saving multi-oscillograms of selected wave objects.
}
\description{
This function launches a Shiny app that allows users to select multiple wave objects, plot their oscillograms, and save the resulting plots as a PNG file. Users can control the maximum duration of the waves to be plotted, set the scale bar, and specify image dimensions.
}
\examples{
multioscillo_app()

}
