% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_app.R
\name{multiplot_app}
\alias{multiplot_app}
\title{Spectrogram, Oscillogram, and Mean Power Spectrum Multiplot Shiny App}
\usage{
multiplot_app()
}
\value{
A Shiny app for visualizing, customizing, and downloading spectrogram and oscillogram plots.
}
\description{
This Shiny app allows users to visualize a combined plot of a spectrogram, mean spectrum, and oscillogram for a selected audio file (Wave object). Users can adjust settings such as noise floor, image dimensions, and background transparency, and can download the resulting plot as a PNG file.
}
\examples{
\dontrun{
if (interactive()) {
  multiplot_app()
}
}
}
