% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_ggplot.R
\name{multiplot_ggplot}
\alias{multiplot_ggplot}
\title{Multiplot including Spectrogram, Mean Power Spectrum, and}
\usage{
multiplot_ggplot(
  wave,
  floor = -35,
  overlap = 95,
  zeropad = 200,
  scale = "dB",
  heights = c(8, 2)
)
}
\arguments{
\item{wave}{A `Wave` object.}

\item{floor}{The noise floor in dBFS.}

\item{overlap}{Window overlap (%).}

\item{zeropad}{Zero padding.}

\item{scale}{Scale for the amplitude. Either "dB" or "Linear".}
}
\value{
A plot with aligned spectrogram, mean spectrum, and oscillogram.
}
\description{
Multiplot including Spectrogram, Mean Power Spectrum, and
}
\examples{
\dontrun{
multiplot_ggplot(coryphoda)
}
}
