% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillogram_plotly.R
\name{oscillogram_plotly}
\alias{oscillogram_plotly}
\title{Interactive Oscillogram}
\usage{
oscillogram_plotly(wave, title = "", line_color = "black")
}
\arguments{
\item{wave}{A wave object containing the audio data to be plotted.}

\item{title}{A string specifying the title of the plot. Default is an
empty string.}

\item{line_color}{A string specifying the color of the waveform line.
Default is 'black'.}
}
\value{
A Plotly object representing the interactive oscillogram.
}
\description{
This function generates an interactive oscillogram plot for a given Wave
object using Plotly.
}
\examples{
\dontrun{
oscillogram_plotly(coryphoda_wave, title = "Coryphoda albidicollis")
}
}
