% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platycleis.R
\docType{data}
\name{platycleis}
\alias{platycleis}
\title{Recording of Platycleis grisea}
\format{
An object of class \code{Wave} of length 1179199.
}
\source{
Original recording by Cesare Brizio.
}
\usage{
platycleis
}
\description{
The calling song of *Platycleis grisea*, a Tettigoniidae
recorded by Cesare Brizio in Italy.
}
\details{
Sampling rate: 250 kHz.

Bit depth: 16 bit.

Duration: ~5 sec.
}
\examples{
\dontrun{
data(platycleis)
summary(platycleis)
spectrum_plotly(platycleis)
}
}
\keyword{datasets}
