% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rthoptera.R
\docType{package}
\name{Rthoptera-package}
\alias{Rthoptera}
\alias{Rthoptera-package}
\title{Rthoptera: Standardized Insect Bioacoustics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package provides functions and Shiny Apps to analyze insect acoustic signals, including preprocessing (e.g., high-pass filter, remove DC offset, down-sampling, etc.). Analytical functions allow the user to automatically extract relevant acoustic features in both spectral and temporal dimensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/naturewaves/Rthoptera}
  \item \url{https://naturewaves.github.io/Rthoptera/}
  \item Report bugs at \url{https://github.com/naturewaves/Rthoptera/issues}
}

}
\keyword{internal}
