% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song_stats_lq.R
\name{song_stats_lq}
\alias{song_stats_lq}
\title{Temporal and Spectral Statistics of LQ (Broad-band) Songs.}
\usage{
song_stats_lq(
  wave,
  specimen_id = "",
  ssmooth = 100,
  peakfinder_ws = 50,
  peakfinder_threshold = 0.005,
  max_peak_gap = 0.01,
  max_train_gap = 0.5,
  motif_seq = TRUE,
  max_motif_gap = 0.8,
  detection_threshold = 0.1,
  norm_env = TRUE,
  db_threshold = 20
)
}
\arguments{
\item{wave}{A `Wave` object.}

\item{specimen_id}{A character string representing the identifier for the
specimen (or recording) being analyzed.}

\item{ssmooth}{Numeric. Smoothing window size (in samples) for the amplitude
envelope.}

\item{peakfinder_ws}{Numeric. Window size (in samples) for the peak detection
algorithm.}

\item{peakfinder_threshold}{Numeric. Amplitude threshold for peak detection,
as a proportion of the maximum amplitude.}

\item{max_peak_gap}{Numeric. Maximum allowed gap (in seconds) between peaks
to consider them part of the same train.}

\item{max_train_gap}{Numeric. Maximum allowed gap (in seconds) between trains
to consider them part of the same motif.}

\item{motif_seq}{Logical. If TRUE, add an first-order aggregation of motifs.
Default = FALSE.}

\item{max_motif_gap}{Numeric. Maximum allowed gap (in seconds) between motifs
to consider them part of the same group.}

\item{detection_threshold}{Numeric. Minimum amplitude for a peak to be
considered valid.}

\item{norm_env}{Logical. Whether to normalize the amplitude envelope between
0 and 1. Defaults to `TRUE`.}

\item{db_threshold}{Numeric. Decibel threshold below the peak amplitude for
calculating low and high frequencies. Defaults to 20 dB.}
}
\value{
A list containing the following components:
  \item{plot}{An interactive `plotly` object showing the waveform envelope,
  detected peaks, trains, and motifs.}
  \item{summary_data}{A tibble with summary statistics for the analyzed
  waveform.}
  \item{motif_seq_data}{A tibble summarizing motif sequences. Only produced
  when motif_seq = TRUE.}
  \item{motif_data}{A tibble summarizing motifs, including motif duration,
  train counts, and spectral properties.}
  \item{train_data}{A tibble detailing detected trains, including start and
  end times, durations, peak frequencies, and bandwidths.}
  \item{peak_data}{A tibble containing detected peaks, including their times,
  periods, and amplitudes.}
  \item{params}{A tibble summarizing input parameters for the function.}
}
\description{
This function calculates temporal and spectral statistics, identifying motifs
and trains in the waveform, detecting peaks, and extracting spectral features
such as peak frequency, low frequency, high frequency, and bandwidth.
}
\examples{
\dontrun{
data(coryphoda)
# Perform analysis on the coryphoda sample
results <- call_stats_lq(
  wave = coryphoda,
  specimen_id = "Coryphoda_Sample",
  ssmooth = 100,
  peakfinder_ws = 50,
  peakfinder_threshold = 0.01,
  max_train_gap = 0.5,
  max_peak_gap = 0.05,
  detection_threshold = 0.2,
  norm_env = TRUE,
  db_threshold = 20
)
results$plot
}
}
