% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_stats.R
\name{spectral_stats}
\alias{spectral_stats}
\title{Calculate Spectral Statistics}
\usage{
spectral_stats(
  wave,
  specimen.id = "Specimen ID",
  total.range = FALSE,
  robust = FALSE,
  ampMax = 0,
  dbth = -20,
  lines = TRUE,
  sound.type = "Call 1",
  temp = NULL,
  hpf = NULL
)
}
\arguments{
\item{wave}{A `Wave` object representing the sound data.}

\item{specimen.id}{Character string for the specimen identifier.}

\item{total.range}{Logical, whether to calculate the full frequency range.}

\item{robust}{Logical, whether to use a robust frequency resolution.}

\item{ampMax}{Numeric, sets the amplitude scaling (1 for relative amplitude, 0 for dB).}

\item{dbth}{Numeric, decibel threshold for bandwidth calculation.}

\item{lines}{Logical, whether to add lines for min, max, and peak frequencies.}

\item{sound.type}{Character string for the type of sound analyzed.}

\item{temp}{Numeric, optional, temperature in degrees Celsius.}

\item{hpf}{Numeric, optional, high-pass filter cutoff frequency in kHz.}
}
\value{
A list containing:
- `data`: A tibble with calculated statistics.
- `plot`: An interactive `plotly` plot with frequency and amplitude traces.
}
\description{
This function calculates various spectral statistics for a given audio signal, including peak frequency, bandwidth, spectral entropy, and spectral flatness. It generates an interactive plot with frequency and amplitude traces and includes options to visualize key statistics with lines on the plot.
}
\examples{
\dontrun{
wave <- readWave("example.wav")
result <- spectral_stats(wave)
result$plot
}

}
