% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_df.R
\name{spectrum_df}
\alias{spectrum_df}
\title{Spectrum Table (Tibble, Data Frame)}
\usage{
spectrum_df(
  wave,
  freq_res = 10,
  fun = "mean",
  wn = "hanning",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{wave}{An object of class \code{Wave} containing the audio data to be
analyzed.}

\item{freq_res}{The frequency resolution to be used for the frequency
spectrum analysis. Use this argument consistently for standardized
measurements across recordings with different sampling rate. Default is
10 Hz per frequency bin.}

\item{fun}{Character string indicating the summary function to be applied
for computing the mean spectrum. Default is \code{'mean'}. Other options include
\code{'median'}, \code{'sd'} (standard deviation), or \code{'var'} (variance).}

\item{wn}{Window filtering function. Choices are:  bartlett, blackman,
flattop, hamming, hanning, or rectangle. Default is \code{hanning}.}

\item{verbose}{Logical. If TRUE, parameter info is printed in the console.}

\item{...}{Other arguments passed to meanspec() from the seewave package.}
}
\value{
A list with two tibbles. The first tibble "spec_df" with two columns:
\code{frequency} and \code{amplitude}, and the second tibble "params_df" contains all
the relevant parameters used to create the spectrum.
The \code{frequency} column contains the frequencies in Hertz (Hz), and
the \code{amplitude} column contains the corresponding amplitude values.
}
\description{
This function generates a data frame containing the frequency and amplitude
of a wave object using the mean spectrum method.
The function leverages the \code{meanspec} function from the \code{seewave} package to
compute the spectrum.
}
\examples{
 \dontrun{
# Assuming `wave` is a loaded Wave object
spec_df <- meanspec_df(wave, from = 0, to = 5, wl = 1024, fun = 'mean')
}
}
