% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_plot.R
\name{spectrum_plot}
\alias{spectrum_plot}
\title{Spectrum Plot with Customizable Options}
\usage{
spectrum_plot(
  wave,
  win.size = NULL,
  ovlp = 50,
  x.breaks = 6,
  y.position = c("left", "right"),
  x.position = c("top", "bottom"),
  flip = FALSE,
  color.db = "grey30",
  color.linear = "black",
  color.carrier = "white",
  color.threshold = "white",
  color.bandwidth = "white",
  fun = c("mean", "median", "var", "sd"),
  wn = c("bartlett", "blackman", "flattop", "hamming", "hanning", "rectangle"),
  show.x.title = TRUE,
  show.y.title = TRUE,
  add.params = FALSE,
  add.summary = TRUE,
  plot.title = "",
  italic.title = FALSE,
  fmin = NULL,
  fmax = NULL,
  linewidth = 1,
  total.bandwidth = FALSE,
  show.lines = FALSE
)
}
\arguments{
\item{wave}{A wave object containing the audio data.}

\item{win.size}{Numeric, the window size for the FFT. If `NULL`, it will be set automatically based on the sample rate.}

\item{ovlp}{Numeric, the overlap between successive windows in percentage. Default is 50.}

\item{x.breaks}{Numeric, the number of breaks on the x-axis (frequency axis). Default is 6.}

\item{y.position}{Character, the position of the y-axis, either "left" or "right". Default is "left".}

\item{x.position}{Character, the position of the x-axis, either "top" or "bottom". Default is "bottom".}

\item{flip}{Logical, if `TRUE`, the plot will be flipped. Default is `FALSE`.}

\item{color.db}{Character, color for the dB amplitude plot. Default is "grey30".}

\item{color.linear}{Character, color for the linear amplitude plot. Default is "black".}

\item{color.carrier}{Character, color for the carrier frequency marker. Default is "white".}

\item{color.threshold}{Character, color for the threshold line. Default is "white".}

\item{color.bandwidth}{Character, color for the bandwidth lines. Default is "white".}

\item{fun}{Character, the summary function to apply to the spectrum, one of "mean", "median", "var", "sd". Default is "mean".}

\item{wn}{Character, the window type for the FFT, one of "bartlett", "blackman", "flattop", "hamming", "hanning", "rectangle". Default is "hanning".}

\item{show.x.title}{Logical, whether to show the x-axis title. Default is `TRUE`.}

\item{show.y.title}{Logical, whether to show the y-axis title. Default is `TRUE`.}

\item{add.params}{Logical, whether to add window parameters as a label on the plot. Default is `FALSE`.}

\item{add.summary}{Logical, whether to add summary statistics as a label on the plot. Default is `TRUE`.}

\item{plot.title}{Character, title of the plot. Default is an empty string.}

\item{italic.title}{Logical, whether to italicize the plot title. Default is `FALSE`.}

\item{fmin}{Numeric, minimum frequency (in kHz) to plot. If `NULL`, the entire frequency range is plotted.}

\item{fmax}{Numeric, maximum frequency (in kHz) to plot. If `NULL`, it defaults to the Nyquist frequency.}

\item{linewidth}{Numeric, the width of the lines in the plot. Default is 1.}

\item{total.bandwidth}{Logical, whether to calculate the total bandwidth, ignoring gaps. Default is `FALSE`.}

\item{show.lines}{Logical, whether to show lines for the carrier frequency and bandwidths. Default is `FALSE`.}
}
\value{
A list containing a ggplot2 object for the spectrum plot.
}
\description{
This function generates a spectrum plot with various customizable options, including window size, color schemes, and the ability to display statistical summaries and frequency markers. The plot can display both dB and linear amplitude scales, and bandwidths based on thresholds can be shown.
}
\examples{
spectrum_plot(wave, win.size = 512, x.breaks = 5, fmin = 1, fmax = 10)
}
