% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_plotly.R
\name{spectrum_plotly}
\alias{spectrum_plotly}
\title{Interactive Spectrum Plot using Plotly}
\usage{
spectrum_plotly(
  wave,
  freq_res = 50,
  db_shade = TRUE,
  ovlp = 0,
  fun = "mean",
  wn = "blackman",
  total_bandwidth = FALSE,
  limit_indices = FALSE,
  plot_title = "",
  italic_title = TRUE,
  fmin = 0,
  fmax = NULL,
  add_params = FALSE,
  add_summary = FALSE,
  x_breaks = 10,
  y_position = "left",
  x_position = "bottom",
  show_x_title = TRUE,
  show_y_title = TRUE,
  color_db = "grey",
  color_linear = "black",
  color_carrier = "white",
  color_threshold = "white",
  color_bandwidth = "white",
  show_lines = FALSE,
  linewidth = 1
)
}
\arguments{
\item{wave}{A wave object containing the audio data.}

\item{freq_res}{Numeric, the frequency resolution in Hz per bin. Default is 50.}

\item{db_shade}{Logical, whether to shade the plot based on decibel amplitude.
Default is \code{TRUE}.}

\item{ovlp}{Numeric, the overlap between successive windows in percentage.
Default is 0 (no overlap).}

\item{fun}{Character, the summary function to apply to the spectrum, one of
"mean", "median", "var", or "sd". Default is "mean".}

\item{wn}{Character, the window type for the FFT, one of "hanning",
"bartlett", "blackman", "flattop", "hamming", or "rectangle". Default is "hanning".}

\item{total_bandwidth}{Logical, whether to calculate the total bandwidth,
ignoring gaps. Default is \code{FALSE}.}

\item{limit_indices}{Logical, whether to limit the spectrum indices to the
frequency range between \code{fmin} and \code{fmax}. Default is \code{FALSE}.}

\item{plot_title}{Character, the title of the plot. Default is an empty string.}

\item{italic_title}{Logical, whether to italicize the plot title. Default is \code{FALSE}.}

\item{fmin}{Numeric, minimum frequency (in kHz) to plot. If \code{NULL},
the entire frequency range is plotted. Default is \code{NULL}.}

\item{fmax}{Numeric, maximum frequency (in kHz) to plot. If \code{NULL}, it
defaults to the Nyquist frequency.}

\item{add_params}{Logical, whether to add window parameters as a label on the plot.
Default is \code{FALSE}.}

\item{add_summary}{Logical, whether to add summary statistics as a label on
the plot. Default is \code{TRUE}.}

\item{x_breaks}{Numeric, the number of breaks on the x-axis (frequency axis). Default is 6.}

\item{y_position}{Character, the position of the y-axis, either "left" or
"right". Default is "left".}

\item{x_position}{Character, the position of the x-axis, either "top" or
"bottom". Default is "bottom".}

\item{show_x_title}{Logical, whether to show the x-axis title. Default is \code{TRUE}.}

\item{show_y_title}{Logical, whether to show the y-axis title. Default is \code{TRUE}.}

\item{color_db}{Character, color for the dB amplitude plot. Default is "grey30".}

\item{color_linear}{Character, color for the linear amplitude plot. Default is "black".}

\item{color_carrier}{Character, color for the carrier frequency marker. Default is "white".}

\item{color_threshold}{Character, color for the threshold line. Default is "white".}

\item{color_bandwidth}{Character, color for the bandwidth lines. Default is "white".}

\item{show_lines}{Logical, whether to show lines for the carrier frequency
and bandwidths. Default is \code{FALSE}.}

\item{linewidth}{Numeric, the width of the lines in the plot. Default is 1.}
}
\value{
A list including: a Plotly object, and a summary table (data frame).
}
\description{
This function generates an interactive spectrum plot using Plotly, offering
customizable options for windowing, overlap, amplitude shading
(dB and linear), and the ability to display summary statistics and parameter
annotations. The function supports both dB and linear scales, bandwidth
calculations, and frequency markers.
}
\examples{
 \dontrun{
# Example usage:
spectrum_plotly(Coryphoda_albidicollis, db_shade = FALSE,
freq_res = 100, show_lines = TRUE)
}
`r lifecycle::badge("stable")`
}
