% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_stats_lq_app.R
\name{temporal_stats_lq_app}
\alias{temporal_stats_lq_app}
\title{Temporal Statistics Low-Quality Shiny App}
\usage{
temporal_stats_lq_app()
}
\value{
A Shiny app for analyzing temporal statistics of acoustic waveforms.
}
\description{
This Shiny app performs low-quality temporal statistics analysis on wave objects. It allows the user to configure parameters such as smoothing window, peak finder thresholds, and gap limits for trains and peaks. The app generates interactive plots and data tables summarizing motifs, trains, and peaks, along with parameter settings.
}
\examples{
 \dontrun{
if (interactive()) {
  temp_stats_lq_app()
}
}
}
