% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_app.R
\name{trim_app}
\alias{trim_app}
\title{Select parts of a wave, and save them in the R environment}
\usage{
trim_app()
}
\value{
One or more Wave objects created from selected parts of the original wave object.
}
\description{
This Shiny app allows users to select parts of a wave object, zoom in or out on the waveform, and save the selected region as a new wave object in the R environment.
}
\examples{
trim_app()

}
