% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_df.R
\name{wave_df}
\alias{wave_df}
\title{Wave Data Frame}
\usage{
wave_df(wave, norm = TRUE)
}
\arguments{
\item{wave}{A Wave object.}

\item{norm}{Logical. If true, the amplitude values are rescaled to a range of [-1,1].}
}
\value{
A tibble data frame.
}
\description{
Wave Data Frame
}
\examples{
wave_df(wave, norm = TRUE)
}
