% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coryphoda.R
\docType{data}
\name{coryphoda}
\alias{coryphoda}
\title{Recording of Coryphoda albidicollis}
\format{
An object of class \code{Wave} of length 84536.
}
\source{
Original recording by Francisco Rivas Fuenzalida.
}
\usage{
coryphoda
}
\description{
A single calling song of *Coryphoda albidicollis*, a Phaneropterinae recorded
by Francisco Rivas in Central Chile.
}
\details{
Sampling rate: 96 kHz.
Temperature: 22.6 (Celsius).
}
\examples{
\dontrun{
data(coryphoda)
summary(coryphoda)
spectrum_plotly(coryphoda)
}
}
\keyword{datasets}
