% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eumodicogryllus.R
\docType{data}
\name{eumodicogryllus}
\alias{eumodicogryllus}
\title{Recording of Eumodicogryllus bordigalensis}
\format{
An object of class \code{Wave} of length 453279.
}
\source{
\url{https://xeno-canto.org/872608}.
}
\usage{
eumodicogryllus
}
\description{
The calling song of *Eumodicogryllus bordigalensis*, a Gryllidae
recorded by Cesare Brizio in Italy.
}
\details{
Sampling rate: 44.1 kHz.

Bit depth: 16 bit.

Duration: ~10 sec.
}
\examples{
\dontrun{
data(eumodicogryllus)
summary(eumodicogryllus)
spectrum_plotly(eumodicogryllus)
}
}
\keyword{datasets}
