% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcidectes.R
\docType{data}
\name{falcidectes}
\alias{falcidectes}
\title{Example Sound File 2}
\format{
An object of class \code{\link[tuneR]{Wave}}. The Wave object contains:
\describe{
  \item{left}{The left audio channel (if stereo).}
  \item{samp.rate}{The sampling rate in Hz.}
  \item{bit}{The bit depth (e.g., 16 for CD quality).}
  \item{duration}{The total duration of the sound file in seconds.}
}
}
\source{
Original recording by Francisco Rivas Fuenzalida.
}
\usage{
falcidectes
}
\description{
The calling song of *Falcidectes sp.*, a Tettigoniidae from Central Chile.
}
\details{
This mono file was recorded at a sampling rate of 96 kHz, 16 bit depth.The recording lasts 4.3 seconds.
The sound file has been imported into an R object using
the \code{\link[tuneR]{readWave}} function from the \pkg{tuneR} package.
}
\examples{
data(coryphoda22C)
summary(coryphoda22C)
spectrum_plotly(coryphoda22C)

}
\keyword{datasets}
