% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillogram_ggplot.R
\name{oscillogram_ggplot}
\alias{oscillogram_ggplot}
\title{Oscillogram with ggplot2}
\usage{
oscillogram_ggplot(wave, y_title = "Relative Amplitude", x_title = "")
}
\arguments{
\item{wave}{A Wave object}

\item{y_title}{Character. A title for the Y-axis.}

\item{x_title}{Character. A title for the X-axis.}
}
\value{
An oscillogram plot. This plot is meant to be used with other
plots with matching the X-axis, therefore it lacks its labels.
}
\description{
Oscillogram with ggplot2
}
\examples{
\dontrun{
oscillgoram_ggplot(coryphoda)
}
}
