% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram_ggplot.R
\name{spectrogram_ggplot}
\alias{spectrogram_ggplot}
\title{Standardized Spectrogram plus Mean Power Spectrum}
\usage{
spectrogram_ggplot(
  wave,
  meanspec = TRUE,
  floor = -35,
  scale = "dB",
  overlap = 95,
  zeropad = 200
)
}
\arguments{
\item{wave}{A wave object.}

\item{meanspec}{Logical. If TRUE, a flipped mean power spectrum is added on the right side of the spectrogram.}

\item{floor}{Noise floor in dBFS.}

\item{scale}{Scale for the amplitude. Either "dB" or "Linear".}

\item{overlap}{Window overlap (%).}

\item{zeropad}{Zero padding.}
}
\value{
A spectrogram plot with automatic window length and optionally a mean power spectrum.
}
\description{
Standardized Spectrogram plus Mean Power Spectrum
}
\examples{
\dontrun{
spectrogram_ggplot()}
}
