% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_pop.R
\name{spectrum_pop}
\alias{spectrum_pop}
\title{Plot a representative spectrum of a population (or subspecies, species, etc.)}
\usage{
spectrum_pop(
  waves,
  binomial = NULL,
  freq_res = 400,
  s_rate_out = 96000,
  fun = "mean",
  wn = "hanning",
  flim = NULL,
  line_color = "green4",
  line_width = 1,
  sd_color = "dodgerblue",
  sd_alpha = 0.5,
  title_x = "Frequency (kHz)",
  title_y = "",
  show_x_labs = TRUE,
  show_y_labs = FALSE,
  show_n = TRUE,
  save_to = NULL,
  save = FALSE,
  width = 960,
  height = 540,
  dpi = 200
)
}
\arguments{
\item{binomial}{Character. The binomial name of the species. Empty by default.}

\item{freq_res}{Numeric. The frequency resolution to be used for the
frequency spectrum analysis. Use this argument consistently for standardized
measurements across recordings with different sampling rate. Default is 10 Hz
per frequency bin.}

\item{s_rate_out}{Sample rate of the output summary PSD, in kHz.}

\item{fun}{Character. Select a function to compute each power spectrum. Options
are "mean", "median", "sd", and "var".}

\item{wn}{Character. Select a window function to for the FFT. Options are
"hanning" (Default), "hamming", bartlett", "blackaman", "flattop", and "rectangular".}

\item{flim}{Numeric list of length 2. Range of the frequency axis (KhZ).}

\item{line_color}{Character. Color for the line representing the mean PSD.}

\item{line_width}{Numeric. Width of the mean line.}

\item{sd_color}{Character. Color for the standard deviation ribbon.}

\item{sd_alpha}{Numeric. Transparency level for the SD ribbon. Default is 0.3 (30\%).}

\item{title_x}{Character. Title for the X-axis. Default is "Frequency (kHz)".}

\item{title_y}{Character. Title for the Y-axis. Empty by default.}

\item{show_x_labs}{Logical. If TRUE, the labels (i.e., numbers) are shown in
the X-axis. Defaults to TRUE.}

\item{show_y_labs}{Logical. If TRUE, the labels (i.e., numbers) are shown in
the Y-axis. Defaults to FALSE}

\item{show_n}{Logical. If TRUE, the number of samples (i.e., recordings) is
shown as "n=" on the top-right of the plot.}

\item{save}{Logical. Whether to save the plot immediately.}

\item{width}{Numeric. Relative width of the output plot. Default is 960.}

\item{height}{Numeric. Relative height of the output plot. Default is 540.}

\item{dpi}{Numeric. Resolution in dots per inch. Default is 200.}

\item{wave}{An object of class \code{Wave} containing the audio data to be
analyzed.}
}
\value{
An averaged and smoothed power spectral density plot representing all
the Waves in the list provided to the \code{waves} argument. The plot shows the
mean and confidence interval (95\% by default) along with the peak frequency,
the scientific name of the species as provided in the \code{binomial} argument,
and the number of samples (Wave objects) from which the mean was calculated.
}
\description{
This function generates an averaged Power Spectral Density (PSD) plot multiple
input Waves (e.g., population, species, etc,).
}
\examples{
 \dontrun{
library(RthopteraSounds)
data(coryphoda)
data(tettigonia)
# Assume the two Waves are two individuals from the same species
waves_list <- c(coryphoda, tettigonia)
population_spectrum(waves = waves_list, freq_res = 50, show_x_labs = TRUE,
title_x = "", show_n = TRUE, binomial = "Coryphoda albidicollis")
}
}
